First draft of proteins, genes, bacteria. Add ant build.
Added first draft of proteins, genes and bacteria.

The proteins contain the information about the Chem field, whether
they are supposed to be expressed or not, their name and threshold
level for expression.

The gene represents a function that takes proteins as input and
produces a protein as output. There are 3 gene implementations:
- gene representing AND function - all proteins must be evaluated to
true in order to output protein,
- gene representing OR function - one protein must be evaluated to
true in order to protein,
- gene representing TRUE function - always outputs protein.

Bacteria holds genes, some represent the fitness functions, while
others represent clasues. It overrides the action function - it checks
all the genes and produces proteins (BSimChemicalfield) if genes
evaluate to true. There is also an interaction function - if two
bacteria collide, there is a possibility of conjugation. That means
that one bacterium will transfer a random clause gene to the other.

WARNING - the code is not tested.

Added an ant build specification that creates a jar file. Removed the
shell scripts.