First version of the application
Redesign of solution.

Proteins now only contain name and if they are supposed to be present.

The genes now take a dictionary when evaluating. They compare their
expected input parameters with the dictionary to determine if all
expected inputs are present. Based on that an output protein is
generated.

The bacterium now holds the F1, F2, F3 genes and the clauses genes
separately. The logic in action is also different - firstly, clauses
genes are evaluated. Then F1 and F2 genes are evaluated. Based on
that, the state of the bacterium is set. The bacterium is colored
based on its state.

The main method collects input data from standard input and prepares
the simulation. It creates proteins and genes and orders them based on
whether they should be expressed. F1, F2 and F3 genes are created
based on input data. Bacteria is created similarly to the original
article. Then the simulation runs.

TODO - nothing implemented with mutation rate yet.