Add new generations with mutation
Implemented multiple generations by allowing bacteria to grow and
replicate. When a bacterium is ready to replicate, it will create a
copy of itself, where there is a mutation rate chance, that a gene
will not be in the child bacterium. The program takes a new input -
max population size, so the simulation does not explode.

Added a new field in bacterium that indicates the generation number.